[tbs_plugins/sma_astro] Use 'long long' instead of 'long' for the fstop counter.
Using just a long meant that the counter would wrap after approximately 24 days
since each iteration was about 1 ms. Now it will wrap after something like
300 million years...

+++ tbs_plugins/sma_astro.c
@@ -415,7 +415,7 @@ void add_nsec(struct timespec *in, struct timespec *out, long long nsec) {
 /* Only this loop should access registers */
 void * fringe_stop(void * tr){
   int j;
-  long i = 0;
+  long long i = 0;
   int result;
 
   int cdelay;
